Fixed tick_t bugs on ARM cpu. On this platform, casing directly  from a double to a 32bit uint will saturate the uint rather than wrap it. Fixes cpu/fps measurement on ARM, and no doubt several things relating to timing generally.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@40185 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/tick.h
@@ -45,10 +45,10 @@
 /* Using a 32-bit type for tick_t prevents wraparound bugs on platforms with 32-bit timers */
 typedef uint32_t tick_t;
 
-#define TICK_TO_MILLI(tick) ((uint32_t) ((double)(tick) * ((double)MILLI_PER_SECOND / TICK_PER_SECOND)))
-#define TICK_TO_MICRO(tick) ((uint32_t) ((double)(tick) * ((double)MICRO_PER_SECOND / TICK_PER_SECOND)))
-#define TICK_TO_NANO(tick)  ((uint64_t) ((double)(tick) * ((double)NANO_PER_SECOND  / TICK_PER_SECOND)))
-#define NANO_TO_TICK(nano)  ((tick_t)   ((double)(nano) / ((double)NANO_PER_SECOND  / TICK_PER_SECOND)))
+#define TICK_TO_MILLI(tick) ((uint32_t) (uint64_t)((double)(tick) * ((double)MILLI_PER_SECOND / TICK_PER_SECOND)))
+#define TICK_TO_MICRO(tick) ((uint32_t) (uint64_t)((double)(tick) * ((double)MICRO_PER_SECOND / TICK_PER_SECOND)))
+#define TICK_TO_NANO(tick)  ((uint64_t) (uint64_t)((double)(tick) * ((double)NANO_PER_SECOND  / TICK_PER_SECOND)))
+#define NANO_TO_TICK(nano)  ((tick_t)   (uint64_t)((double)(nano) / ((double)NANO_PER_SECOND  / TICK_PER_SECOND)))
 
 extern void tick_init(void);
